Create a logger class to centralize logging code.
