+++ Phoenix/Filters/ExceptionLogAttribute.cs
@@ -1,50 +1,37 @@
-﻿using System;
+﻿using Phoenix.Services;
+using System;
 using System.IO;
+using System.Text;
 using System.Web.Hosting;
 using System.Web.Mvc;
 
 namespace Phoenix.Filters
 {
+    /// <summary>
+    /// Log any uncaught exceptions in controller classes.
+    /// </summary>
     public class ExceptionLogAttribute : FilterAttribute, IExceptionFilter
     {
         public void OnException(ExceptionContext filterContext)
         {
-            string today = DateTime.Today.ToShortDateString().Replace("/", "_");
+            LoggerService log = new LoggerService();
+            StringBuilder message  = new StringBuilder();
 
-            // Msdn doesn't have enough documentation on DateTime, so I used:
-            // http://www.c-sharpcorner.com/uploadfile/mahesh/working-with-datetime-using-C-Sharp/
-            // to figure out the type of timestamp format I wanted
-            string timestamp = DateTime.Today.ToString("G");
-
-            string folderPath = "\\Logs\\";
-            Directory.CreateDirectory(HostingEnvironment.MapPath(folderPath));
-
-            var stream = File.AppendText(HostingEnvironment.MapPath(folderPath + today + ".log"));
-
-            stream.Write(timestamp);
-            stream.Write(" --- ");
-            stream.Write("[ERROR]");
-            stream.Write(" ");
             // Unsolicited information: The user agent string is a mess.
             // Here is an article to get you up to date:
             // http://webaim.org/blog/user-agent-string-history/
             // FUN
-            stream.WriteLine(filterContext.HttpContext.Request.UserAgent);
-            stream.Write(filterContext.HttpContext.Request.UserHostAddress);
-            stream.Write(" ");
-            stream.Write(filterContext.HttpContext.Request.HttpMethod.ToString());
-            stream.Write(" ");
-            stream.WriteLine(filterContext.HttpContext.Request.Url.ToString());
-
-            stream.WriteLine("Exception: " + filterContext.Exception.Message);
-
-            stream.WriteLine("Inner Exception: " + filterContext.Exception.InnerException);
-
-            stream.WriteLine();
-            stream.WriteLine();
+            message.AppendLine("User-Agent: " + filterContext.HttpContext.Request.UserAgent);
+            message.Append(filterContext.HttpContext.Request.UserHostAddress);
+            message.Append(" ");
+            message.Append(filterContext.HttpContext.Request.HttpMethod.ToString());
+            message.Append(" ");
+            message.AppendLine(filterContext.HttpContext.Request.Url.ToString());
+            message.AppendLine("Exception: " + filterContext.Exception.Message);
+            message.AppendLine("Inner Exception: " + filterContext.Exception.InnerException);
 
 
-            stream.Dispose();
+            log.Error(message.ToString());
         }
     }
 }
