+++ Phoenix/Phoenix.csproj
@@ -202,6 +202,7 @@
     </Compile>
     <Compile Include="Services\AdminDashboardService.cs" />
     <Compile Include="Services\DashboardService.cs" />
+    <Compile Include="Services\LoggerService.cs" />
     <Compile Include="Services\LoginService.cs" />
     <Compile Include="Services\RciComponentReassignService.cs" />
     <Compile Include="Services\RciInputService.cs" />
